Updating floats to doubles
gives better image resolution at high zooms

+++ include/mandelbrot.h
@@ -5,12 +5,12 @@ typedef struct mandelbrot_fractal *Mandelbrot;
 typedef struct mandelbrot_fractal {
 
     // The coordinates of the bottom left corner
-    float x1;
-    float y1;
+    double x1;
+    double y1;
 
     // The coordinates of the top right corner
-    float x2;
-    float y2;
+    double x2;
+    double y2;
 
     // The size of the area in pixels
     // Allows us to calculate the complex
@@ -30,15 +30,15 @@ typedef struct mandelbrot_fractal {
 } Mandelbrot_Data;
 
 // Create the Mandelbrot Data struct and populate it with data
-Mandelbrot brot_create(int pixWidth, int pixHeight, int repeats, float x1, float y1, float x2, float y2);
+Mandelbrot brot_create(int pixWidth, int pixHeight, int repeats, double x1, double y1, double x2, double y2);
 
-Mandelbrot brot_recreate(Mandelbrot brot, float x1, float y1, float x2, float y2);
+Mandelbrot brot_recreate(Mandelbrot brot, double x1, double y1, double x2, double y2);
 
 Mandelbrot brot_calculate(Mandelbrot brot);
 
 int brot_pixel_coords(Mandelbrot brot, int xPos, int yPos);
 
-int brot_calc_escape(float xPos, float yPos, float limit, int depth);
+int brot_calc_escape(double xPos, double yPos, double limit, int depth);
 
 // Cleanup the Mandelbrot data struct and free all the assigned memory
 void brot_cleanup(Mandelbrot brot);
